# api/swagger-spec/swagger.yml
swagger: "2.0"
info:
  title: Workflow Manager
  version: 2.0.0
produces:
  - application/json
consumes:
  - application/json
schemes:
  - http
paths:
...
  /v3/doctor/{uuid}:
    parameters:
      - $ref: "#/parameters/UUID"
    get:
      operationId: getDoctorInfo
      summary: "get the specified doctor report as per UUID"
      security:
        - basic: []
      responses:
        200:
          description: doctor get response
          schema:
            $ref: "#/definitions/doctorInfo"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
    post:
      operationId: publishDoctorInfo
      summary: "publish doctor info to Workflow Manager API"
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/doctorInfo"
      responses:
        200:
          description: publish doctorInfo response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
...
parameters:
...
  UUID:
    name: uuid
    type: string
    in: path
    description: A universal Id to represent a sepcific request or report
    required: true
securityDefinitions:
  basic:
    type: basic
definitions:
  cluster:
    type: object
    required:
      - id
      - components
    properties:
      id:
        type: string
        minLength: 1
      firstSeen:
        type: string
        format: date-time
      lastSeen:
        type: string
        format: date-time
      components:
        type: array
        items:
          $ref: "#/definitions/componentVersion"
...
  k8sResource:
    type: object
    required:
      - data
    properties:
      data:
        type: object
  namespace:
    type: object
    required:
      - name
      - daemonSets
      - deployments
      - events
      - pods
      - replicaSets
      - replicationControllers
      - services
    properties:
      name:
        type: string
      daemonSets:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      deployments:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      events:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      pods:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      replicaSets:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      replicationControllers:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      services:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
  doctorInfo:
    type: object
    required:
      - workflow
      - nodes
      - namespaces
    properties:
      workflow:
        $ref: "#/definitions/cluster"
      nodes:
        type: array
        items:
          $ref: "#/definitions/k8sResource"
      namespaces:
        type: array
        items:
          $ref: "#/definitions/namespace"
  componentVersion:
    type: object
    properties:
      component:
        $ref: "#/definitions/component"
      version:
        $ref: "#/definitions/version"
      updateAvailable:
        type: string
  component:
    type: object
    required:
      - name
    properties:
      name:
        type: string
        minLength: 1
      description:
        type: string
      type:
        type: string
  version:
    type: object
    properties:
      train:
        type: string
        minLength: 1
      version:
        type: string
        minLength: 1
      released:
        type: string
        minLength: 1
      data:
        $ref: "#/definitions/versionData"
  versionData:
    type: object
    properties:
      description:
        type: string
        minLength: 1
      fixes:
        type: string
        minLength: 1
      image:
        type: string
  error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string